+++ src/main/java/ru/transport/Transport.java
@@ -2,17 +2,17 @@ package ru.transport;
 
 public class Transport {
     protected String name;
-    protected int price;
+    protected double price;
 
     public String getName() {
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 
-    public Transport(String name, int price) {
+    public Transport(String name, double price) {
         this.name = name;
         this.price = price;
     }
